refactor: extract isHook to util
